REFACTOR(`StateContainer`): moved to subfolder
