+++ OnlineMinion.Web/GlobalUsings.cs
@@ -1,0 +1,3 @@
+// Global using directives
+
+global using OnlineMinion.Web.State;
