+++ OnlineMinion.Web/OnlineMinion.Web.csproj
@@ -28,10 +28,10 @@
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\OnlineMinion.Common.Utilities\OnlineMinion.Common.Utilities.csproj"/>
-        <ProjectReference Include="..\OnlineMinion.Common\OnlineMinion.Common.csproj"/>
-        <ProjectReference Include="..\OnlineMinion.Contracts\OnlineMinion.Contracts.csproj"/>
-        <ProjectReference Include="..\OnlineMinion.RestApi.Client\OnlineMinion.RestApi.Client.csproj"/>
+        <ProjectReference Include="..\OnlineMinion.Common.Utilities\OnlineMinion.Common.Utilities.csproj" />
+        <ProjectReference Include="..\OnlineMinion.Common\OnlineMinion.Common.csproj" />
+        <ProjectReference Include="..\OnlineMinion.Contracts\OnlineMinion.Contracts.csproj" />
+        <ProjectReference Include="..\OnlineMinion.RestApi.Client\OnlineMinion.RestApi.Client.csproj" />
     </ItemGroup>
 
 </Project>
